Перламутровые пуговицы. Третья проверка
